Rename container/presentational components to AppNav
